LPS-58119 extract interface
