+++ modules/portal-store/portal-store-s3/src/com/liferay/portal/store/s3/S3KeyTransformer.java
@@ -14,86 +14,27 @@
 
 package com.liferay.portal.store.s3;
 
-import com.liferay.portal.kernel.util.CharPool;
-import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.kernel.util.StringPool;
-
 /**
- * @author Edward C. Han
+ * @author Michael C. Han
  */
-public class S3KeyTransformer {
+public interface S3KeyTransformer {
 
 	public String getDirectoryKey(
-		long companyId, long repositoryId, String folderName) {
-
-		return getFileKey(companyId, repositoryId, folderName);
-	}
+		long companyId, long repositoryId, String folderName);
 
 	public String getFileKey(
-		long companyId, long repositoryId, String fileName) {
+		long companyId, long repositoryId, String fileName);
 
-		StringBundler sb = new StringBundler(4);
-
-		sb.append(companyId);
-		sb.append(StringPool.SLASH);
-		sb.append(repositoryId);
-		sb.append(getNormalizedFileName(fileName));
-
-		return sb.toString();
-	}
-
-	public String getFileName(String key) {
-
-		// Convert /${companyId}/${repositoryId}/${dirName}/${fileName}
-		// /${versionLabel} to ${dirName}/${fileName}
-
-		int x = key.indexOf(CharPool.SLASH);
-
-		x = key.indexOf(CharPool.SLASH, x + 1);
-
-		int y = key.lastIndexOf(CharPool.SLASH);
-
-		return key.substring(x + 1, y);
-	}
+	public String getFileName(String key);
 
 	public String getFileVersionKey(
 		long companyId, long repositoryId, String fileName,
-		String versionLabel) {
+		String versionLabel);
 
-		StringBundler sb = new StringBundler(6);
+	public String getNormalizedFileName(String fileName);
 
-		sb.append(companyId);
-		sb.append(StringPool.SLASH);
-		sb.append(repositoryId);
-		sb.append(getNormalizedFileName(fileName));
-		sb.append(StringPool.SLASH);
-		sb.append(versionLabel);
+	public String getRepositoryKey(long companyId, long repositoryId);
 
-		return sb.toString();
-	}
-
-	public String getNormalizedFileName(String fileName) {
-		String normalizedFileName = fileName;
-
-		if (!fileName.startsWith(StringPool.SLASH)) {
-			normalizedFileName = StringPool.SLASH + normalizedFileName;
-		}
-
-		if (fileName.endsWith(StringPool.SLASH)) {
-			normalizedFileName = normalizedFileName.substring(
-				0, normalizedFileName.length() - 1);
-		}
-
-		return normalizedFileName;
-	}
-
-	public String getRepositoryKey(long companyId, long repositoryId) {
-		return companyId + StringPool.SLASH + repositoryId;
-	}
-
-	public String moveKey(String key, String oldPrefix, String newPrefix) {
-		String name = key.substring(oldPrefix.length());
-		return newPrefix.concat(name);
-	}
+	public String moveKey(String key, String oldPrefix, String newPrefix);
 
 }
